add: Api 응답 핸들링 추가
